move some things to a dedicated cc
