+++ src/s2geography.h
@@ -11,6 +11,7 @@
 #include "s2geography/distance.h"
 #include "s2geography/geoarrow.h"
 #include "s2geography/geography.h"
+#include "s2geography/geography_interface.h"
 #include "s2geography/index.h"
 #include "s2geography/linear-referencing.h"
 #include "s2geography/predicates.h"
